Fracturing in Path object's parametric function
Currently, the parametric function does not fracture the path when the number of computed points is greater than 'max_points'.  I believe it is because max_points (on line 2235) takes the max value of np (the number of computed points) and max_points//2.  This should really be the *min* of the two, since if np > max_points, there should be fracturing.  Please let me know if I overlooked something.